rename module pypop -> PyPop to avoid namespace collisions removing more kruft update instructions
