+++ setup.py
@@ -36,13 +36,9 @@
 
 import sys, os, string
 
-#from distutils.core import setup, Extension
 from setuptools import setup
 from setuptools.extension import Extension
-#from distutils.file_util import copy_file
-#from distutils.sysconfig import PREFIX, get_config_vars, get_config_var
 from sysconfig import _PREFIX, get_config_vars, get_config_var
-#from distutils.command.build_ext import build_ext
 
 # Override the overzealous use of _FORTIFY_SOURCE CFLAGS flags that
 # are in /usr/lib/python2.4/config/Makefile used on Fedora Core 4
@@ -50,59 +46,6 @@ from sysconfig import _PREFIX, get_config_vars, get_config_var
 # http://mail.python.org/pipermail/distutils-sig/2002-December/003123.html
 cv = get_config_var("OPT")
 cv = cv.replace("-D_FORTIFY_SOURCE=2", "-D_FORTIFY_SOURCE=1")
-
-# override implementation of swig_sources method in standard build_ext
-# class, so we can change the way SWIG is called by Python's default
-# configuration of distutils
-# class my_build_ext(build_ext):
-
-#     def swig_sources (self, sources, extension=None):
-
-#         """Walk the list of source files in 'sources', looking for SWIG
-#         interface (.i) files.  Run SWIG on all that are found, and
-#         return a modified 'sources' list with SWIG source files replaced
-#         by the generated C (or C++) files.
-#         """
-
-#         new_sources = []
-#         swig_sources = []
-#         swig_targets = {}
-
-#         # XXX this drops generated C/C++ files into the source tree, which
-#         # is fine for developers who want to distribute the generated
-#         # source -- but there should be an option to put SWIG output in
-#         # the temp dir.
-
-#         if self.swig_cpp:
-#             target_ext = '.cpp'
-#         else:
-#             target_ext = '.c'
-
-#         for source in sources:
-#             (base, ext) = os.path.splitext(source)
-#             if ext == ".i":             # SWIG interface file
-#                 new_sources.append(base + target_ext)
-#                 swig_sources.append(source)
-#                 swig_targets[source] = new_sources[-1]
-#             else:
-#                 new_sources.append(source)
-
-#         if not swig_sources:
-#             return new_sources
-
-#         swig = self.find_swig()
-#         swig_cmd = [swig, "-python", "-ISWIG"]
-#         if self.swig_cpp:
-#             swig_cmd.append("-c++")
-
-#         for source in swig_sources:
-#             target = swig_targets[source]
-#             self.announce("swigging %s to %s" % (source, target))
-#             self.spawn(swig_cmd + ["-o", target, source])
-
-#         return new_sources
-
-    # swig_sources ()
 
 # distutils doesn't currently have an explicit way of setting CFLAGS,
 # it takes CFLAGS from the environment variable of the same name, so
@@ -117,22 +60,6 @@ if "LIBRARY_PATH" in os.environ:
     library_dirs += os.environ["LIBRARY_PATH"].rstrip(os.pathsep).split(os.pathsep)
 if "CPATH" in os.environ:
     include_dirs += os.environ["CPATH"].rstrip(os.pathsep).split(os.pathsep)
-
-print include_dirs
-
-# flag to determine whether or not we are using the CVS version
-if os.path.isdir("CVS"):
-    cvs_version=1
-else:
-    cvs_version=0
-
-# flag to determine whether we are generating a distribution version
-if os.environ.has_key('DISTRIB') and \
-   os.environ['DISTRIB'] == 'true':
-    distrib_version=1
-else:
-    distrib_version=0
-
 
 # define each extension
 ext_Emhaplofreq = Extension("_Emhaplofreqmodule",
@@ -241,10 +168,10 @@ ext_Gthwe.depends=['SWIG/typemap.i', 'gthwe/func.h', 'gthwe/hwe.h']
 # default list of extensions to build
 extensions = [ext_Emhaplofreq, ext_EWSlatkinExact, ext_Pvalue, ext_Gthwe]
 
-# don't include HWE
+# don't include HWE yet
 # extensions.append(ext_HweEnum)
 
-from pypop import __version__, __pkgname__
+from PyPop import __version__, __pkgname__
 
 # data files to install
 data_file_paths = ['config.ini']
@@ -263,17 +190,13 @@ particularly large-scale multilocus genotype data""",
        maintainer_email = "alexl@cal.berkeley.edu",
        license = "GNU GPL",
        platforms = ["GNU/Linux", "Windows", "MacOS"],
-       packages = ["pypop"],
-       #install_requires = [
+       packages = ["PyPop"],
+       # install_requires = [
        #  'Numeric',
        #  'libxml2-python'
        #  ],
-       #py_modules = ["Arlequin", "HardyWeinberg", "Utils", "Haplo",
-       #              "Homozygosity", "ParseFile", "Filter", "Main",
-       #              "DataTypes", "GUIApp", "RandomBinning", "Meta"],
        scripts= ['pypop.py', 'popmeta.py'],
        data_files=[('share/pypop', data_file_paths)],
-       #cmdclass = {'build_ext': my_build_ext,},  # compile SWIG module
        ext_modules=extensions
        )
